requirements.txt
transformers==4.36.0
torch==2.1.0
datasets==2.14.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
huggingface-hub>=0.19.3
accelerate==0.24.0
sentencepiece==0.1.99
google-generativeai>=0.3.0
requests==2.31.0
